requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
email-validator==2.1.0
phonenumbers==8.13.27
Werkzeug==3.0.1